+++ docs/conf.py
@@ -12,8 +12,8 @@ sys.path.insert(0, os.path.abspath("../src"))
 project = "OpenPKPD"
 author = "OpenPKPD contributors"
 copyright = "2025, OpenPKPD contributors"
-release = "0.3.3"
-version = "0.3.3"
+release = "0.3.4"
+version = "0.3.4"
 
 # ---------------------------------------------------------------------------
 # Extensions